feat: Add support for PHP 8.2 (#212)
## Small Fixes due to PHP8.2 upgrade:
- Internal PHP `ZendHashmap` changed the way that we access to the bucket ([commit](https://github.com/php/php-src/commit/90b7bde61507cee1c6b37f153909d72f5b203b8c#diff-dca4782c62f10b418ea7193e4641e688ffd7a9c97869f2b1a3a96e28da5653aaL371))
- Internal PHP `ArrayIterator` doesn't support parent `get_iterator` ([commit](https://github.com/php/php-src/commit/15bbf6f3377e6446e88300b9c5383179091fe270#diff-bc002b59b592b4d44c13d898f9dc3ca60b1b1c8505a51a2a3557dfde51dfeb61L273))

+++ src/zend/ex.rs
@@ -213,7 +213,7 @@ impl ExecuteData {
     #[doc(hidden)]
     unsafe fn zend_call_var_num(&self, n: isize) -> *mut Zval {
         let ptr = self as *const Self as *mut Zval;
-        ptr.offset(Self::zend_call_frame_slot() + n as isize)
+        ptr.offset(Self::zend_call_frame_slot() + n)
     }
 
     /// Translation of macro `ZEND_CALL_FRAME_SLOT`
